+++ src/containers/Write/Funnel/NewsCollection.tsx
