+++ src/containers/Write/Funnel/WriteTargetInfo.tsx
