+++ src/pages/Write.tsx
@@ -1,17 +1,17 @@
 import Header from '@/components/Header'
 import { HEADER_TITLE, SELECT_TARGET_BUTTONS, WRITE_STEPS } from '@/constants/write'
-import MessagesCollection from '@/containers/Write/MessagesCollection'
-import NewsCollection from '@/containers/Write/NewsCollection'
-import SearchOffice from '@/containers/Write/SearchOffice'
-import WriteMessage from '@/containers/Write/WriteMessage'
+import MessagesCollection from '@/containers/Write/Funnel/MessagesCollection'
+import NewsCollection from '@/containers/Write/Funnel/NewsCollection'
+import SearchOffice from '@/containers/Write/Funnel/SearchOffice'
+import WriteMessage from '@/containers/Write/Funnel/WriteMessage'
 import WriteMessageHeader from '@/containers/Write/WriteMessageHeader'
-import WriteTargetInfo from '@/containers/Write/WriteTargetInfo'
+import WriteTargetInfo from '@/containers/Write/Funnel/WriteTargetInfo'
 import useBodyBackgroundColor from '@/hooks/useBodyBackgroundColor'
 import useFunnel from '@/hooks/useFunnel'
 import useWriteMessageStore from '@/stores/writeMessageStore'
 import { useEffect } from 'react'
-import SelectIsSpecific from '@/containers/Write/SelectIsSpecific'
-import SelectHeroType from '@/containers/Write/SelectHeroType'
+import SelectIsSpecific from '@/containers/Write/Funnel/SelectIsSpecific'
+import SelectHeroType from '@/containers/Write/Funnel/SelectHeroType'
 
 export default function Write() {
   const clearTargetType = useWriteMessageStore((state) => state.clearTargetType)
